tests/wpt/web-platform-tests/html/semantics/scripting-1/the-script-element/execution-timing/029.html

<!DOCTYPE html>
<html><head>
        <title> scheduler: javascript: URL in HREF</title>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="testlib/testlib.js"></script>
</head>
<body>

        <div id="log">FAILED (This TC requires JavaScript enabled)</div>
        <p><a href="javascript:log('JS URL')"></a></p>
        <script>log('inline script #1');
                if(document.links[0].click){
                        document.links[0].click();
                }else{
                        var evt = document.createEvent("MouseEvents");
                        evt.initMouseEvent("click", true, true, window,
                        0, 0, 0, 0, 0, false, false, false, false, 0, null);
                        document.links[0].dispatchEvent(evt);
                }
                log('end script #1');
        </script>

        <script type="text/javascript">
        log( 'inline script #2' );
        var t = async_test()

        function test() {
                assert_any(assert_array_equals, eventOrder, [
                  ['inline script #1', 'end script #1', 'JS URL', 'inline script #2'],
                  ['inline script #1', 'end script #1', 'inline script #2', 'JS URL']]);
                t.done();
        }
        // Note: The JS URL should be executed in a task, queued by the task doing the navigation.
        // Therefore, when the load event fires, the JS URL will not have executed yet.
        // A time-out is added to ensure the JS URL has executed by the time we run the assertions.
        onload = t.step_func(function() { t.step_timeout(test, 1000) });
        </script>

</body></html>
