LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/execution-timing/029-expected.txt

FAILED (This TC requires JavaScript enabled)

FAIL  scheduler: javascript: URL in HREF assert_array_equals: property 1, expected "end script #1" but got "JS URL"

assert_array_equals: property 1, expected "end script #1" but got "JS URL"
